refactor: ApiMember 삭제 및 중복 토큰 검증 로직 삭제
