+++ src/main/java/org/ject/support/common/security/config/WebConfig.java
@@ -2,7 +2,7 @@ package org.ject.support.common.security.config;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.ject.support.common.security.ApiMemberArgumentResolver;
+import org.ject.support.common.security.AuthenticatedMemberIdResolver;
 import org.ject.support.common.security.jwt.JwtTokenProvider;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
@@ -12,10 +12,8 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 @RequiredArgsConstructor
 public class WebConfig implements WebMvcConfigurer {
 
-    private final JwtTokenProvider jwtTokenProvider;
-
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new ApiMemberArgumentResolver(jwtTokenProvider));
+        resolvers.add(new AuthenticatedMemberIdResolver());
     }
 }
